TelegramAutomation/telegram-automation/requirements.txt
flask==3.0.0
flask-cors==4.0.0
asyncpg==0.29.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
